requirements.txt
Flask
gunicorn
discord.py